[refactor] 未登録グループのAPIフックを分離、関連する処理を整理
